Move destinations function and give it token
